Round gas depth properly
The D in MOD, EAD, END, and EADD stands for "depth" and
as such these should be mm in int rather than double.

The intermediate fn2 and fhe2, however, as intermediate
value should not be rounded to an integer.

The upshot of this is a litle more numerical stability.
It should lead to more stable values in TestProfile
when run on architectures with different floating
point precision.

+++ core/profile.h
@@ -58,7 +58,7 @@ struct plot_data {
 	pressure_t o2sensor[3]; //for rebreathers with up to 3 PO2 sensors
 	pressure_t o2setpoint;
 	pressure_t scr_OC_pO2;
-	double mod, ead, end, eadd;
+	int mod, ead, end, eadd;
 	velocity_t velocity;
 	int speed;
 	// stats over 9 minute window:
